Refacto créerEspècesGroupéesParImpact Extrait de la logique qui regroupe toutes les activités par code
