+++ scripts/front-end/actions/créerEspècesGroupéesParImpact.js
@@ -1,4 +1,3 @@
-import {chargerActivitésMéthodesTransports} from './dossier.js'
 
 /** @import { ActivitéMenançante, DescriptionMenacesEspèces, FauneNonOiseauAtteinte, FloreAtteinte, OiseauAtteint } from '../../types/especes.d.ts' */
 
@@ -94,53 +93,10 @@ let activitéVersDonnéesSecondaires = new Map([
 /**
  * 
  * @param {DescriptionMenacesEspèces} espècesImpactées
- * @returns {Promise<EspècesParActivité[]>}
+ * @param {Map<ActivitéMenançante['Code'], ActivitéMenançante>} activitéByCode
+ * @returns {EspècesParActivité[]}
  */
-export async function créerEspècesGroupéesParImpact(espècesImpactées) {
-
-    /** @type {Map<ActivitéMenançante['Code'], ActivitéMenançante>} */
-    const activitéByCode = await chargerActivitésMéthodesTransports()
-        .then(({activités}) => {
-            // Rajouter les activités spécifiques Pitchou
-            // Les activités sont standardisées à l'échelle européenne
-            // https://dd.eionet.europa.eu/schemas/habides-2.0/derogations.xsd (type 'activitiesType')
-            // Pour les besoins de Pitchou, nous rajoutons des activités 
-            // Nous essayons d'utiliser des identifiants qui ne collisionnerons pas avec le futur
-
-            const activité4 = activités.oiseau.get('4')
-            if(!activité4){
-                throw Error(`Activité 4 manquante`)
-            }
-
-            /** @type {Map<ActivitéMenançante['Code'], ActivitéMenançante>} */
-            // @ts-ignore
-            const activitésAdditionnelles = new Map([
-                {
-                    ...activité4,
-                    Code: '4-1-pitchou-aires',
-                    "étiquette affichée": `Destruction d’aires de repos ou reproduction`
-                },
-                {
-                    ...activité4,
-                    Code: '4-2-pitchou-nids',
-                    "étiquette affichée": `Destruction de nids`
-                },
-                {
-                    ...activité4,
-                    Code: '4-3-pitchou-œufs',
-                    "étiquette affichée": `Destruction d'œufs`
-                }
-            ].map(a => [a.Code, a]))
-
-            return new Map([
-                ...activités.oiseau, 
-                ...activitésAdditionnelles, 
-                ...activités['faune non-oiseau'], 
-                ...activités.flore
-            ])
-        })
-
-
+export function créerEspècesGroupéesParImpact(espècesImpactées, activitéByCode) {
 
     /** @type {Map<ActivitéMenançante | undefined, EspèceImpactéeSimplifiée[]>} */
     const _espècesImpactéesParActivité = new Map()
