+++ scripts/front-end/actions/dossier.js
@@ -228,7 +228,7 @@ export function actMetTransArraysToMapBundle(activitésBrutes, méthodesBrutes, 
 }
 
 /**
- * 
+ * @remark à déplacer dans un fichier activitésMéthodesTransports ?
  * @returns {Promise<NonNullable<PitchouState['activitésMéthodesTransports']>>}
  */
 export async function chargerActivitésMéthodesTransports(){
@@ -271,4 +271,57 @@ export async function espècesImpactéesDepuisFichierOdsArrayBuffer(fichierArray
         transports
     )
 
+}
+
+/**
+ * Crée un Map unifié de toutes les activités menaçantes indexées par leur code.
+ * 
+ * @returns {Promise<Map<ActivitéMenançante['Code'], ActivitéMenançante>>}
+ * @remark à déplacer dans un fichier activitésMéthodesTransports ?
+ * @see {@link https://dd.eionet.europa.eu/schemas/habides-2.0/derogations.xsd}
+ */
+export async function getActivitéByCodeP() {
+    const activitéByCode = await chargerActivitésMéthodesTransports()
+        .then(({activités}) => {
+
+            // Rajouter les activités spécifiques Pitchou
+            // Les activités sont standardisées à l'échelle européenne
+            // https://dd.eionet.europa.eu/schemas/habides-2.0/derogations.xsd (type 'activitiesType')
+            // Pour les besoins de Pitchou, nous rajoutons des activités 
+            // Nous essayons d'utiliser des identifiants qui ne collisionnerons pas avec le futur
+
+             const activité4 = activités.oiseau.get('4')
+             if(!activité4){
+                 throw Error(`Activité 4 manquante`)
+             }
+
+             /** @type {Map<ActivitéMenançante['Code'], ActivitéMenançante>} */
+            //@ts-ignore
+             const activitésAdditionnelles = new Map([
+                 {
+                     ...activité4,
+                     Code: '4-1-pitchou-aires',
+                     "étiquette affichée": `Destruction d’aires de repos ou reproduction`
+                 },
+                 {
+                     ...activité4,
+                     Code: '4-2-pitchou-nids',
+                     "étiquette affichée": `Destruction de nids`
+                 },
+                 {
+                     ...activité4,
+                     Code: '4-3-pitchou-œufs',
+                     "étiquette affichée": `Destruction d'œufs`
+                 }
+             ].map(a => [a.Code, a]))
+
+             return new Map([
+                 ...activités.oiseau, 
+                 ...activitésAdditionnelles, 
+                 ...activités['faune non-oiseau'], 
+                 ...activités.flore
+             ])
+         })
+
+    return activitéByCode
 }
